List Zenodo files

List all files available in a given Zenodo record. Requires already having a Zenodo account set up and

zenodo_list(
  concept_doi,
  token = Sys.getenv("zenodo_token"),
  sandbox = FALSE,
  as_datatable = TRUE,
  verbose = TRUE
)

Source

zen4R vignette zen4R vignette

Arguments

concept_doi

Concept DOI of the Zenodo record.

token

Zenodo Personal access token.

sandbox

Whether to use the Zenodo or Zenodo Sandbox API.

as_datatable

Return file info as a data.table.

verbose

Print messages.

Value

Nested list of Zenodo file info.

Examples

if (FALSE) { # \dontrun{
zen_dat <- zenodo_list(token="<token>",
                       concept_doi="10.5281/zenodo.7062237")
} # }